Switch to Xenomai 3.1
Make 3.1 our distro's default version, removing the previous "stable"
version. The latter is currently equivalent to master.

The idea of setting the default to a release is providing the user a
well-defined version, rather than stable of the day. We will introduce
a stable head kas fragment once needed.

With "stable" gone, we also need to update the kernel version selection
magic. It is only needed for "XENOMAI_3_0" now.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>